chore(tests): moved tests to their own directories
